+++ src/components/Sidebar.tsx
@@ -76,12 +76,6 @@ const Sidebar: FC<SidebarProps> = ({ close, isOpen }) => {
                     href="https://www.github.com/ld5ehom"
                     target="_blank"
                 />
-
-                {/* eslint-disable-next-line @next/next/no-img-element */}
-                <img
-                    src="https://hits.seeyoufarm.com/api/count/incr/badge.svg?url=https%3A%2F%2Fnextjs-blog-chatbot.vercel.app&count_bg=%2379C83D&title_bg=%23555555&icon=&icon_color=%23E7E7E7&title=Visitor&edge_flat=false"
-                    alt="Visitor Badge"
-                />
             </div>
         </div>
     );
