+++ src/pages/admin/index.tsx
@@ -1,3 +1,4 @@
+import Button from "@/components/Button";
 import Input from "@/components/Input";
 import { createClient } from "@/utils/supabase/client";
 import { UserResponse } from "@supabase/supabase-js";
@@ -12,7 +13,7 @@ export default function Admin() {
     const emailRef = useRef<HTMLInputElement>(null); // Reference for the email input field. (이메일 입력 필드에 대한 참조입니다.)
     const passwordRef = useRef<HTMLInputElement>(null); // Reference for the password input field. (비밀번호 입력 필드에 대한 참조입니다.)
 
-    // Handle form submission for login.)
+    // Handle form submission for login.
     const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
         e.preventDefault(); // Prevent the default form submission behavior. (기본 폼 제출 동작을 방지합니다.)
         const response = await supabase.auth.signInWithPassword({
@@ -39,7 +40,7 @@ export default function Admin() {
 
     return (
         // Main container for the admin page. (관리자 페이지의 메인 컨테이너입니다.)
-        <div className="container mx-auto flex flex-col px-4 pb-20 pt-12">
+        <div className="container flex flex-col pb-20 pt-12">
             {!!userResponse?.data.user ? (
                 // Render this section if the user is logged in. (사용자가 로그인된 경우 이 섹션을 렌더링합니다.)
                 <div className="flex flex-col gap-2">
@@ -49,25 +50,20 @@ export default function Admin() {
                     </div>
 
                     {/* Button to navigate to the post creation page. (글 작성 페이지로 이동하는 버튼입니다.) */}
-                    <button
-                        type="button"
-                        className="w-full rounded-md bg-gray-800 py-2 text-white"
-                        onClick={() => router.push("/write")}
-                    >
+                    <Button type="button" onClick={() => router.push("/write")}>
                         Create Post
-                    </button>
+                    </Button>
 
                     {/* Button to log out and navigate to the home page. (로그아웃하고 홈 페이지로 이동하는 버튼입니다.) */}
-                    <button
+                    <Button
                         type="button"
-                        className="w-full rounded-md bg-gray-800 py-2 text-white"
                         onClick={() => {
                             supabase.auth.signOut(); // Log out the user. (사용자를 로그아웃합니다.)
                             router.push("/"); // Redirect to the home page. (홈 페이지로 리다이렉트합니다.)
                         }}
                     >
                         Log Out
-                    </button>
+                    </Button>
                 </div>
             ) : (
                 // Render this section if the user is not logged in. (사용자가 로그인하지 않은 경우 이 섹션을 렌더링합니다.)
