+++ src/pages/write/index.tsx
@@ -1,7 +1,8 @@
+import Button from "@/components/Button";
 import Input from "@/components/Input";
 import { MarkdownEditor } from "@/components/Markdown";
+import { useCategories, useTags } from "@/utils/hooks";
 import { createClient } from "@/utils/supabase/client";
-import { useQuery } from "@tanstack/react-query";
 import { useRouter } from "next/router";
 import { FormEvent, useRef, useState } from "react";
 import ReactSelect from "react-select/creatable";
@@ -14,22 +15,8 @@ export default function Write() {
     const titleRef = useRef<HTMLInputElement>(null);
     const fileRef = useRef<HTMLInputElement>(null);
 
-    const { data: existingCategories } = useQuery({
-        queryKey: ["categories"],
-        queryFn: async () => {
-            const { data } = await supabase.from("Post").select("category");
-            return Array.from(new Set(data?.map((d) => d.category)));
-        },
-    });
-    const { data: existingTags } = useQuery({
-        queryKey: ["tags"],
-        queryFn: async () => {
-            const { data } = await supabase.from("Post").select("tags");
-            return Array.from(
-                new Set(data?.flatMap((d) => JSON.parse(d.tags)))
-            );
-        },
-    });
+    const { data: existingCategories } = useCategories();
+    const { data: existingTags } = useTags();
 
     const [category, setCategory] = useState("");
     const [tags, setTags] = useState("");
@@ -71,7 +58,7 @@ export default function Write() {
     };
 
     return (
-        <div className="container mx-auto flex flex-col px-4 pb-20 pt-12">
+        <div className="container flex flex-col pb-20 pt-12">
             <h1 className="mb-8 text-2xl font-medium">New Post</h1>
             <form onSubmit={handleSubmit}>
                 <div className="flex flex-col gap-3">
@@ -118,12 +105,9 @@ export default function Write() {
                 </div>
 
                 {/* Submit button (제출 버튼) */}
-                <button
-                    type="submit"
-                    className="mt-4 w-full rounded-md bg-uclaBlue py-2 text-white"
-                >
+                <Button type="submit" className="mt-4">
                     Submit
-                </button>
+                </Button>
             </form>
         </div>
     );
